+++ templates/ios/XcodeProjectTemplate/projectName/Scenes/containerName/cellNameCollectionViewCell.swift.hbs
@@ -29,8 +29,13 @@ class {{classPrefix}}CollectionViewCell: BindableCollectionViewCell<[TreeElement
     }
 
     override func bind(data: [TreeElement]) {
+        guard let viewConfig = self.viewConfig else { return }
         for (uid, name) in self.viewIdMap {
-            let targetName = self.trimIfNeeded(name)
+            // you need to exclude dynamic classes itself.
+            // because all `data`'s treeElements are belongings of the dynamic classes,
+            // so both `name` and `name within the data` will never match on `getProperty(name)`.
+            // that's why I need to exclude here.
+            let targetName = viewConfig.trimClassPrefix(name, classPrefix: "{{treeName}}")
             guard
                 let view = self.viewConfig?.getView(uid),
                 let props = data.getProperty(targetName) else { continue }
@@ -47,23 +52,5 @@ class {{classPrefix}}CollectionViewCell: BindableCollectionViewCell<[TreeElement
             }
 
         }
-    }
-}
-
-extension {{classPrefix}}CollectionViewCell {
-    func trimIfNeeded(_ name: String) -> String {
-
-        // you need to exclude dynamic classes itself.
-        // because all `data`'s treeElements are belongings of the dynamic classes,
-        // so both `name` and `name within the data` will never match on `getProperty(name)`.
-        // that's why I need to exclude here.
-        var targetName = name
-        if let matchedRange = name.range(of: "^{{treeName}}", options: .regularExpression) {
-            let removed = targetName.replacingCharacters(in: matchedRange, with: "")
-            targetName = removed.prefix(1) == "."
-                ? String(removed[removed.index(after: removed.startIndex)..<removed.endIndex])
-                : removed
-        }
-        return targetName
     }
 }
