+++ templates/ios/XcodeProjectTemplate/projectName/Views/Common/Rect+Codable.swift
@@ -1,0 +1,20 @@
+import Foundation
+import UIKit
+
+struct Rect: Codable {
+    var x: CGFloat
+    var y: CGFloat
+    var width: CGFloat
+    var height: CGFloat
+
+    var rect: CGRect {
+        return CGRect(x: x, y: y, width: width, height: height)
+    }
+
+    init(rect: CGRect) {
+        x = rect.origin.x
+        y = rect.origin.y
+        width = rect.size.width
+        height = rect.size.height
+    }
+}
